refactor: add test for model
