+++ app/domain/entity/model.go
